[Refactor] 면접 평가표 생성 페이지 개선
